Separate UserList, add checks before calling onChange
